requirements.txt
xmltodict
python-dotenv
fastapi
uvicorn
selenium
chromedriver_binary